+++ tests/posts/crud.rs
@@ -6,23 +6,21 @@ use super::utils;
 use unanimitylibrary::database::models::prelude::Post;
 use unanimitylibrary::lib::seeds::posts::seed_test_posts;
 
-#[test]
-fn read_all_posts() {
-    let route = "/api/v1/posts";
+const POSTS_ROUTE: &'static str = "/api/v1/posts";
 
+#[test]
+fn read_all_posts_while_logged_in() {
     // clean database
     let client = init::clean_client();
+    init::seed();
     let conn = init::database_connection();
-
-    // create random posts
-    seed_test_posts(&conn);
 
     // login
     let (user, password) = init::get_user(true);
     let auth_token_header = init::login(&user.email, &password);
 
     // perform request
-    let req = client.get(route).header(auth_token_header);
+    let req = client.get(POSTS_ROUTE).header(auth_token_header);
     let mut response = req.dispatch();
 
     //check the answer is Ok
@@ -33,9 +31,107 @@ fn read_all_posts() {
     let posts: Vec<Post> = serde_json::from_str(&data).unwrap();
     // we want a total of 6 post, the 5 normals & the locked one
     assert_eq!(posts.len(), 6);
+
+    // assert we can see the locked one
+    assert_eq!(posts.iter().filter(|p| p.locked).count(), 1);
+
+    // assert we cant see the hidden one
+    assert_eq!(posts.iter().filter(|p| p.hidden).count(), 0);
+
+    // assert we cant see a deleted
+    assert_eq!(posts.iter().filter(|p| p.deleted).count(), 0);
+
+    // the others must be "normal"
+    assert_eq!(
+        posts
+            .iter()
+            .filter(|p| !p.hidden && !p.hidden && !p.locked)
+            .count(),
+        5
+    );
 }
 
-// read admin all posts
+#[test]
+fn read_all_posts_while_logged_in_as_admin() {
+    // clean database
+    let client = init::clean_client();
+    init::seed();
+    let conn = init::database_connection();
+
+    // login
+    let auth_token_header = init::login_admin();
+
+    // perform request
+    let req = client.get(POSTS_ROUTE).header(auth_token_header);
+    let mut response = req.dispatch();
+
+    //check the answer is Ok
+    assert_eq!(response.status(), Status::Ok);
+
+    // check the answer data is what we wanted
+    let data = response.body_string().unwrap();
+    let posts: Vec<Post> = serde_json::from_str(&data).unwrap();
+    // we want a total of 7 post, the 5 normals, the locked one & the hidden one
+    assert_eq!(posts.len(), 7);
+
+    // assert we can see the locked one
+    assert_eq!(posts.iter().filter(|p| p.locked).count(), 1);
+
+    // assert we can see the hidden one
+    assert_eq!(posts.iter().filter(|p| p.hidden).count(), 1);
+
+    // assert we cant see a deleted
+    assert_eq!(posts.iter().filter(|p| p.deleted).count(), 0);
+
+    // the others must be "normal"
+    assert_eq!(
+        posts
+            .iter()
+            .filter(|p| !p.hidden && !p.hidden && !p.locked)
+            .count(),
+        5
+    );
+}
+
+#[test]
+fn read_all_posts_without_being_logged_in() {
+    // clean database
+    let client = init::clean_client();
+    init::seed();
+    let conn = init::database_connection();
+
+    // perform request
+    let req = client.get(POSTS_ROUTE);
+    let mut response = req.dispatch();
+
+    //check the answer is Ok
+    assert_eq!(response.status(), Status::Ok);
+
+    // check the answer data is what we wanted
+    let data = response.body_string().unwrap();
+    let posts: Vec<Post> = serde_json::from_str(&data).unwrap();
+    // we want a total of 6 post, the 5 normals, the hidden & the locked one
+    assert_eq!(posts.len(), 6);
+
+    // assert we can see the locked one
+    assert_eq!(posts.iter().filter(|p| p.locked).count(), 1);
+
+    // assert we can see the hidden one
+    assert_eq!(posts.iter().filter(|p| p.hidden).count(), 0);
+
+    // assert we cant see a deleted
+    assert_eq!(posts.iter().filter(|p| p.deleted).count(), 0);
+
+    // the others must be "normal"
+    assert_eq!(
+        posts
+            .iter()
+            .filter(|p| !p.hidden && !p.hidden && !p.locked)
+            .count(),
+        5
+    );
+}
+
 // read all posts with a search term
 // read all posts related to a tag
 // read all posts with a sorting criteria